chaoxing/requirements.txt
requests
pyaes
beautifulsoup4
lxml
argparse
loguru
celery
flask
fonttools
openai
pycryptodome
ddddocr==1.5.6; python_version < "3.13"
PySide6>=6.7.0